requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-bigquery
langchain
langchain-google-vertexai
langchain-community
